522: v1.3: add user
