+++ app/server.py
@@ -36,6 +36,7 @@ from routes.check_list import check_list
 from routes.logs import logs
 from routes.lti import lti
 from routes.login import login
+from routes.user import user_blueprint
 
 from server_consts import UPLOAD_FOLDER, ALLOWED_EXTENSIONS, DOCUMENT_TYPES, TABLE_COLUMNS, URL_DOMEN
 
@@ -55,6 +56,7 @@ app.register_blueprint(check_list, url_prefix='/check_list')
 app.register_blueprint(logs, url_prefix='/logs')
 app.register_blueprint(lti, url_prefix='/lti')
 app.register_blueprint(login, url_prefix='/login')
+app.register_blueprint(user_blueprint, url_prefix='/user')
 
 app.logger.addHandler(get_logging_stdout_handler())
 app.logger.propagate = False
@@ -77,15 +79,6 @@ def signup():
     elif request.method == "POST":
         u = user.signup(request.json)
         return u.username if u is not None and login_user(u, remember=True) else ""
-
-
-@app.route("/user", methods=["GET", "PUT", "DELETE"])
-@login_required
-def interact():
-    if request.method == "GET":
-        return user.logout()
-    elif request.method == "PUT":
-        return user.edit(request.json)
 
 
 # Main chapters req handlers:
